--- tauri/src-tauri/src/application/mod.rs
@@ -1,1 +1,0 @@
-pub mod services;
