--- tauri/src-tauri/src/application/services/mod.rs
@@ -1,2 +1,0 @@
-mod rendezvous_service;
-pub use rendezvous_service::RendezvousService;
