+++ tauri/src-tauri/src/handlers/p2p_handlers.rs
@@ -1,10 +1,10 @@
-use crate::application::services::RendezvousService;
 use crate::types::CryptoResponse;
 use crate::types::InitiateFirstConnectionInput;
 use log::info;
 use osvauld_core::models::p2p::ConnectionType;
 use osvauld_services::UserService;
 use p2p_service::P2PService;
+use rendezvous_client::rendezvous_service::RendezvousService;
 use std::sync::Arc;
 use sys_locale::get_locale;
 use tauri::State;
